r2039/sofa-dev : FIX: should fix windows compilation and warnings

+++ modules/sofa/component/topology/SparseGridTopology.h
@@ -224,15 +224,15 @@ protected:
     SparseGridTopology* _finerSparseGrid; ///< an eventual finer sparse grid that can be used to built this coarser sparse grid
     SparseGridTopology* _coarserSparseGrid; ///< an eventual coarser sparse grid
 
-    void setVoxel(unsigned int index, unsigned char value)
+    void setVoxel(int index, unsigned char value)
     {
         if (value)
         {
-            (*dataVoxels.beginEdit())[index>>3] |= (int) pow(2.0f, index%8);
+            (*dataVoxels.beginEdit())[index>>3] |= (int) pow(2.0f, (int)(index%8));
         }
         else
         {
-            const unsigned int i = index%8;
+            const int i = index%8;
             const int mask = (int) pow(2.0f, i);
             if (((*dataVoxels.beginEdit())[index>>3]&mask)>>i) (*dataVoxels.beginEdit())[index>>3] -= mask;
         }
@@ -240,7 +240,7 @@ protected:
 
     bool getVoxel(unsigned int index) const
     {
-        const unsigned int i = index%8;
+        const int i = index%8;
         unsigned char c = dataVoxels.getValue()[index>>3];
         return ((c&((int)(pow(2.0f, i)))) >> i) == 1;
     };
